allow compiler transformation in Portractor codemod
