+++ test/__fixtures__/v7/source/compilerFunctions.js
@@ -1,4 +1,6 @@
 exports.config = {
+    framework: 'jasmine',
+
     cucumberOpts: {
         requireModule: [
             () => {
