+++ test/__fixtures__/v7/source/spec.js
@@ -2,6 +2,8 @@ import { Given, When, Then } from 'cucumber';
 const { Given2, When2, Then2 } = require('cucumber');
 
 exports.config = {
+    framework: 'jasmine',
+
     mochaOpts: {
         ui: 'bdd',
         timeout: 5000,
